refactor: add 파일 구조 변경
1. 도메인.game 패키지 추가로 인한 파일 이동
